// Koora-API/src/app/Models/players-dto.ts
export class PlayersDTO {
    id: number;
    firstname: string;
    lastName: string;
    image: string;
    poste: string;
    nombre: number;
    nationality: string;
    // tslint:disable-next-line:variable-name
    date_Nais: Date;
    about: string;
    vote: string;
    // tslint:disable-next-line:variable-name
    id_team: number;
    file: File;
}